+++ src/reconnector.ts
@@ -2,7 +2,29 @@
 
 const backoff = require('backoff-web');
 
+import { JXT } from 'stanza';
+import Client from './client';
+
+const CXFR_NAMESPACE = 'urn:xmpp:cxfr';
+export const Definition = [
+  {
+    path: 'iq.cxfr',
+    element: 'query',
+    namespace: CXFR_NAMESPACE,
+    fields: {
+      domain: JXT.childText(null, 'domain'),
+      server: JXT.childText(null, 'server')
+    }
+  }
+];
+
 export default class Reconnector {
+  client: Client;
+  backoff: any;
+  _hasConnected = false;
+  _cleanupReconnect: any;
+  _backoffActive = false;
+
   constructor (client) {
     this.client = client;
     this.backoff = backoff.exponential({
@@ -14,14 +36,16 @@ export default class Reconnector {
 
     this.backoff.failAfter(10);
 
-    this.backoff.on('ready', (number, delay) => {
+    this.backoff.on('ready', (num, delay) => {
       if (this.client.connected) {
         this.client.logger.debug('Backoff ready, client already connected');
         return;
       }
+
+      // TODO: this.client._stanzaio.transport.conn may not exist in v12
       if (this.client._stanzaio.transport &&
-          this.client._stanzaio.transport.conn) {
-        const conn = this.client._stanzaio.transport.conn;
+          (this.client._stanzaio.transport as any).conn) {
+        const conn = (this.client._stanzaio.transport as any).conn;
         if (conn.readyState <= 1) {
           if (conn.readyState === 1) {
             this.client.logger.debug('Backoff ready, client not connected, but has websocket open');
@@ -39,7 +63,7 @@ export default class Reconnector {
     });
 
     this.backoff.on('fail', () => {
-      this.hardReconnect();
+      return this.hardReconnect();
     });
 
     // self bound methods so we can clean up the handlers
@@ -54,7 +78,7 @@ export default class Reconnector {
       const temporaryFailure = err && err.condition === 'temporary-auth-failure';
       const channelExpired = this._hasConnected && err && err.condition === 'not-authorized';
       if (channelExpired) {
-        this.hardReconnect();
+        return this.hardReconnect();
       } else if (temporaryFailure) {
         this.client.logger.info('Temporary auth failure, continuing reconnect attempts');
       } else {
@@ -64,18 +88,7 @@ export default class Reconnector {
     });
 
     const stanzaio = this.client._stanzaio;
-    stanzaio.disco.addFeature('urn:xmpp:cxfr');
-    const CxfrStanza = stanzaio.stanzas.define({
-      name: 'cxfr',
-      namespace: 'urn:xmpp:cxfr',
-      tags: ['cxfr'],
-      element: 'query',
-      fields: {
-        domain: stanzaio.stanzas.utils.textSub('urn:xmpp:cxfr', 'domain'),
-        server: stanzaio.stanzas.utils.textSub('urn:xmpp:cxfr', 'server')
-      }
-    });
-    stanzaio.stanzas.extendIQ(CxfrStanza);
+    stanzaio.stanzas.define(Definition);
 
     this.client.on('iq:set:cxfr', () => {
       // After 10 minutes, reconnect automatically
@@ -83,18 +96,18 @@ export default class Reconnector {
       // If no `pending` response received from app, proceed with reconnect
       const failureTimeout = setTimeout(() => {
         clearTimeout(timeout);
-        this.client.reconnect();
+        return this.client.reconnect();
       }, 1000);
       // send request to app to reconnect. app can say `pending` to allow for max 1 hour
       // delay in reconnect, and/or `done` to proceed with reconnect immediately
-      this.client._stanzaio.emit('requestReconnect', (response) => {
+      this.client._stanzaio.emit('requestReconnect' as any, (response) => {
         if (response.pending === true) {
           clearTimeout(failureTimeout);
         }
         if (response.done) {
           clearTimeout(failureTimeout);
           clearTimeout(timeout);
-          this.client.reconnect();
+          return this.client.reconnect();
         }
       });
     });
@@ -106,7 +119,7 @@ export default class Reconnector {
     this.client.logger.info('Attempting to reconnect with new channel.');
     this._cleanupReconnect();
     this._hasConnected = false;
-    this.client.connect();
+    return this.client.connect();
   }
 
   cleanupReconnect () {