+++ src/utils.ts
@@ -9,20 +9,20 @@ function buildUri (host, path, version = 'v2') {
   return `https://api.${host}/api/${version}/${path}`;
 }
 
-export function requestApi (path, { method, data, host, version, contentType, authToken, logger }) {
-  let response = request[method](buildUri(host, path, version))
-    .use(reqlogger(logger, data))
-    .set('Authorization', `Bearer ${authToken}`)
-    .type(contentType || 'json');
+export function requestApi (path, opts: { method?, data?, host?, version?, contentType?, authToken?, logger? }) {
+  let response = request[opts.method](buildUri(opts.host, path, opts.version))
+    .use(reqlogger.bind(this, opts.logger, opts.data))
+    .set('Authorization', `Bearer ${opts.authToken}`)
+    .type(opts.contentType || 'json');
 
-  return response.send(data); // trigger request
+  return response.send(opts.data); // trigger request
 }
 
-export function timeoutPromise (fn, timeoutMs, msg, details) {
+export function timeoutPromise (fn, timeoutMs, msg, details?) {
   return new Promise(function (resolve, reject) {
     const timeout = setTimeout(function () {
       const err = new Error(`Timeout: ${msg}`);
-      err.details = details;
+      (err as any).details = details;
       reject(err);
     }, timeoutMs);
     const done = function () {